==> 01_data.ipynb <==
import pandas as pd
import numpy as np
import warnings
from tqdm import tqdm
import os
warnings.simplefilter('ignore')
tqdm.pandas()

pd.set_option('max_columns', None)
pd.set_option('max_rows', None)
pd.set_option('max_colwidth', 200)

# 存款信息

df_cunkuan = pd.DataFrame()
for f in os.listdir('raw_data/x_train/cunkuan_train'):
    mon = int((f.split('.')[0]).split('_')[-1].replace('m', ''))
    m = (mon-1) % 3+1
    q = (mon-1)//3+1

    tmp = pd.read_csv('raw_data/x_train/cunkuan_train/'+f)
    tmp['m'] = m
    tmp['q'] = q

    df_cunkuan = df_cunkuan.append(tmp)

for f in os.listdir('raw_data/x_test/cunkuan_test'):
    mon = int((f.split('.')[0]).split('_')[-1].replace('m', ''))
    m = (mon-1) % 3+1
    q = (mon-1)//3+1

    tmp = pd.read_csv('raw_data/x_test/cunkuan_test/'+f)
    tmp['m'] = m
    tmp['q'] = q

    df_cunkuan = df_cunkuan.append(tmp)

df_cunkuan.reset_index(drop=True)
df_cunkuan.head()

os.makedirs('data', exist_ok=True)
df_cunkuan.to_pickle('data/cunkuan.pkl')

# 有效用户

df_user_Q1 = pd.read_csv('raw_data/x_test/cust_avli_Q1.csv')
df_user_Q3 = pd.read_csv('raw_data/x_train/cust_avli_Q3.csv')
df_user_Q4 = pd.read_csv('raw_data/x_train/cust_avli_Q4.csv')

df_user_label_Q3 = pd.read_csv('raw_data/y_train_3/y_Q3_3.csv')
df_user_label_Q4 = pd.read_csv('raw_data/y_train_3/y_Q4_3.csv')

df_user_Q3 = df_user_Q3.merge(df_user_label_Q3, how='left')
df_user_Q4 = df_user_Q4.merge(df_user_label_Q4, how='left')

df_user_Q1['q'] = 1
df_user_Q3['q'] = 3
df_user_Q4['q'] = 4

df_user = pd.concat([df_user_Q1, df_user_Q3, df_user_Q4]
                    ).reset_index(drop=True)
df_user.head()

os.makedirs('data', exist_ok=True)
df_user.to_pickle('data/user.pkl')

# 额外用户

# Q1 使用 info 里的全量用户
df_user_Q1 = pd.read_csv('raw_data/x_test/cust_info_q1.csv')
df_user_Q1 = df_user_Q1[['cust_no']]
# Q3 使用 info 里的全量用户
df_user_Q3 = pd.read_csv('raw_data/x_train/cust_avli_Q3.csv')
# Q4 使用 avli 里的有效用户
df_user_Q4 = pd.read_csv('raw_data/x_train/cust_avli_Q4.csv')

df_user_label_Q3 = pd.read_csv('raw_data/y_train_3/y_Q3_3.csv')
df_user_label_Q4 = pd.read_csv('raw_data/y_train_3/y_Q4_3.csv')

df_user_Q3 = df_user_Q3.merge(df_user_label_Q3, how='left')
df_user_Q4 = df_user_Q4.merge(df_user_label_Q4, how='left')

df_user_Q1['q'] = 1
df_user_Q3['q'] = 3
df_user_Q4['q'] = 4

df_user = pd.concat([df_user_Q1, df_user_Q3, df_user_Q4]
                    ).reset_index(drop=True)
df_user.head()

os.makedirs('data', exist_ok=True)
df_user.to_pickle('data/user_all.pkl')

#  用户基本信息

df_user_info_Q1 = pd.read_csv('raw_data/x_test/cust_info_q1.csv')
df_user_info_Q3 = pd.read_csv('raw_data/x_train/cust_info_q3.csv')
df_user_info_Q4 = pd.read_csv('raw_data/x_train/cust_info_q4.csv')

df_user_info_Q1['q'] = 1
df_user_info_Q3['q'] = 3
df_user_info_Q4['q'] = 4

df_user_info = pd.concat(
    [df_user_info_Q1, df_user_info_Q3, df_user_info_Q4]).reset_index(drop=True)

del df_user_info['I7'], df_user_info['I9'], df_user_info['I12']

df_user_info.head()

os.makedirs('data', exist_ok=True)
df_user_info.to_pickle('data/cust_info.pkl')

# 用户重大历史数据

df_user_event_Q1 = pd.read_csv(
    'raw_data/x_test/big_event_test/big_event_Q1.csv')
df_user_event_Q3 = pd.read_csv(
    'raw_data/x_train/big_event_train/big_event_Q3.csv')
df_user_event_Q4 = pd.read_csv(
    'raw_data/x_train/big_event_train/big_event_Q4.csv')

df_user_event_Q1['q'] = 1
df_user_event_Q3['q'] = 3
df_user_event_Q4['q'] = 4

df_user_event = pd.concat(
    [df_user_event_Q1, df_user_event_Q3, df_user_event_Q4]).reset_index(drop=True)
df_user_event.head()

os.makedirs('data', exist_ok=True)
df_user_event.to_pickle('data/event.pkl')

# 用户行为数据

df_behavior = pd.DataFrame()
for f in os.listdir('raw_data/x_train/behavior_train'):
    mon = int((f.split('.')[0]).split('_')[-1].replace('m', ''))
    m = (mon-1) % 3+1
    q = (mon-1)//3+1

    tmp = pd.read_csv('raw_data/x_train/behavior_train/'+f)
    tmp['m'] = m
    tmp['q'] = q

    df_behavior = df_behavior.append(tmp)

for f in os.listdir('raw_data/x_test/behavior_test'):
    mon = int((f.split('.')[0]).split('_')[-1].replace('m', ''))
    m = (mon-1) % 3+1
    q = (mon-1)//3+1

    tmp = pd.read_csv('raw_data/x_test/behavior_test/'+f)
    tmp['m'] = m
    tmp['q'] = q

    df_behavior = df_behavior.append(tmp)

df_behavior.reset_index(drop=True)
df_behavior.head()

os.makedirs('data', exist_ok=True)
df_behavior.to_pickle('data/behavior.pkl')

# 用户资产数据

df_aum = pd.DataFrame()
for f in os.listdir('raw_data/x_train/aum_train'):
    mon = int((f.split('.')[0]).split('_')[-1].replace('m', ''))
    m = (mon-1) % 3+1
    q = (mon-1)//3+1

    tmp = pd.read_csv('raw_data/x_train/aum_train/'+f)
    tmp['m'] = m
    tmp['q'] = q

    df_aum = df_aum.append(tmp)

for f in os.listdir('raw_data/x_test/aum_test'):
    mon = int((f.split('.')[0]).split('_')[-1].replace('m', ''))
    m = (mon-1) % 3+1
    q = (mon-1)//3+1

    tmp = pd.read_csv('raw_data/x_test/aum_test/'+f)
    tmp['m'] = m
    tmp['q'] = q

    df_aum = df_aum.append(tmp)

df_aum.reset_index(drop=True)
df_aum.head()

os.makedirs('data', exist_ok=True)
df_aum.to_pickle('data/aum.pkl')
